mac80211: ath11k: add HACK patch to fix failing sysupgrade
Add HACK patch to fix failing sysupgrade on any device that have ath11k
wifi card.

Due to some BUG, some packet in the tx ring are never "complated" and
moved to the tx completion ring. This cause the related idr of the
packet to never be freed and num_tx_pending never decreated to 0. This
cause the flush function to timeout and sysupgrade to fail as it takes
too much time to terminates the process.

Workaround this on the driver side instead of adding an hack to the .sh
file to make it easier to drop and track in the future.

The workaround is quite simple, when tx_flush is called and the function
timeouts every ring is put under lock and idr freed of the stuck
packets.

THIS IS NOT A FIX BUT A WORKAROUND FOR AN ANNOYING PROBLEM.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>

Reference:
https://github.com/Ansuel/openwrt/commit/1e12cd1ee8ee03e4f28249eb13d81a4d808a74ae